refactor: rename and relocate files part 2.
