+++ src/components/Directory/index.tsx
@@ -1,10 +1,10 @@
 "use client";
 import { useState } from "react";
-import DirectoryOrg from "./DirectoryOrg/directory-org";
+import DirectoryOrg from "./DirectoryOrg";
 import mockDirectoryData from "../../app/mock/mock-directory";
 import { Industry } from "../../app/types";
 
-import Filter from "./Filter/Filter";
+import Filter from "./Filter";
 // import SearchBar from "./search-bar";
 
 export default function Directory() {
